integrated toggle, updated findchildnode
