+++ SatiscribeFrontEnd/components/large blocks/FullTranscriptBlock/Tags and Labels/transcriptTagsLabels.js
@@ -5,7 +5,7 @@ import logos from '@/styles/Logos.module.css'
 import contentblock from '@/styles/components/contentblocks.module.css'
 import tags from '@/styles/components/large blocks/tags.module.css'
 
-function TranscriptTags({type, name}) {
+function TranscriptTags({type, name, disabled}) {
     return (
         <>
         { (type == 'tags' && name =='highlight') ? (
@@ -29,7 +29,12 @@ function TranscriptTags({type, name}) {
                 <div className={logos.evensmallerclickable} style={{ backgroundImage: `url("/icons/Cancellation.png")`, zIndex: 1}}></div>
             </div>
             
-        ) : (type =='labels') ? (
+        ) : (type =='labels' && disabled) ? (
+            <div className={`${tags.labelsTagDisabled} ${flexi.flexRowNoGap} ${flexi.alignCenter}  ${flexi.justifyCenter}`}>
+                <h6 className={`${tags.transcribeCleaningTagText}`} style={{color:`var(--Final_Gray)`, width:'75%'}}>{name}</h6>
+                <div className={logos.evensmallerclickable} style={{ backgroundImage: `url("/iconsFinalGray/Trash.png")`, zIndex: 1}}></div>
+            </div>
+        ):(type =='labels') ? (
             <div className={`${tags.labelsTag} ${flexi.flexRowNoGap} ${flexi.alignCenter}  ${flexi.justifyCenter}`}>
                 <h6 className={`${tags.transcribeCleaningTagText}`} style={{color:`var(--Final_Gray)`, width:'75%'}}>{name}</h6>
                 <div className={logos.evensmallerclickable} style={{ backgroundImage: `url("/iconsFinalGray/Trash.png")`, zIndex: 1}}></div>
