+++ SatiscribeFrontEnd/styles/components/large blocks/tags.module.css
@@ -56,3 +56,11 @@ h6.transcribeCleaningTagText {
     border-radius: 5px;
     box-shadow: 2px 2px 4px rgba(0, 0, 0, 0.3);
 }
+
+.labelsTagDisabled{
+    background-color: var(--Final_White_50);
+    width: 14vw;
+    height: 4.5vh;
+    border-radius: 5px;
+    box-shadow: 2px 2px 4px rgba(0, 0, 0, 0.3);
+}
